+++ src/calliope/__init__.py
@@ -1,6 +1,6 @@
 """Public packaging of Calliope."""
 
-from rich import pretty
+import importlib
 
 from calliope import examples
 from calliope._version import __version__
@@ -8,7 +8,12 @@ from calliope.attrdict import AttrDict
 from calliope.model import Model, read_netcdf, read_yaml
 from calliope.util.logging import set_log_verbosity
 
-pretty.install(max_depth=1)
+try:
+    from rich import pretty
+
+    pretty.install(max_depth=1)
+except ModuleNotFoundError:
+    pass
 
 __title__ = "Calliope"
 __author__ = "Calliope contributors listed in AUTHORS"
